Rename imagenet.py to img_224_cls.py
