+++ benchmarker/modules/problems/resnet50/data.py
@@ -1,1 +1,1 @@
-from benchmarker.util.data.synthetic.imagenet import get_data
+from benchmarker.util.data.synthetic.img_224_cls import get_data
