+++ benchmarker/util/data/synthetic/img_224_cls.py
@@ -1,13 +1,13 @@
-"""Generate synthetic data for ImageNet type problems"""
+"""Generate synthetic data for 224x224 images"""
 
 
 from .helpers import gen_classification_data
 
 
 def get_data(params):
-    """Generate synthetic data for ImageNet type problems. Set
-    `params["size"]` appropriately. Import this function in the
-    `data.py` of the problem, so it can be called by `INeuralNet`.
+    """Generate synthetic 224x224 images. Set `params["size"]`
+    appropriately. Import this function in the `data.py` of the
+    problem, so it can be called by `INeuralNet`.
 
     """
